requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.8.2
textblob==0.18.0
google-genai==1.5.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1

httpx>=0.28.1
